Documentation/Makefile: create tarballs for the man pages and html files
[jc: rewrote by stealing from what I run to update html and
 man branches automatically]

+++ Documentation/install-webdoc.sh
@@ -4,12 +4,16 @@ T="$1"
 
 for h in *.html *.txt howto/*.txt howto/*.html
 do
-	diff -u -I'Last updated [0-9][0-9]-[A-Z][a-z][a-z]-' "$T/$h" "$h" || {
+	if test -f "$T/$h" &&
+	   diff -u -I'Last updated [0-9][0-9]-[A-Z][a-z][a-z]-' "$T/$h" "$h"
+	then
+		:; # up to date
+	else
 		echo >&2 "# install $h $T/$h"
 		rm -f "$T/$h"
 		mkdir -p `dirname "$T/$h"`
 		cp "$h" "$T/$h"
-	}
+	fi
 done
 strip_leading=`echo "$T/" | sed -e 's|.|.|g'`
 for th in "$T"/*.html "$T"/*.txt "$T"/howto/*.txt "$T"/howto/*.html
